refactor duplicated code
