+++ components/internal/typescript-base-compiler/index.js
@@ -1,0 +1,1 @@
+export { typescript } from './typescript-base'
